+++ modules/groovy/ui/src/sap/commerce/toolset/groovy/actionSystem/GroovyWebContextActionGroup.kt
@@ -1,0 +1,66 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package sap.commerce.toolset.groovy.actionSystem
+
+import com.intellij.openapi.actionSystem.*
+import com.intellij.openapi.actionSystem.ex.ActionUtil
+import sap.commerce.toolset.HybrisIcons
+import sap.commerce.toolset.groovy.GroovyExecConstants
+import sap.commerce.toolset.groovy.editor.groovyExecContextSettings
+import sap.commerce.toolset.groovy.editor.groovyWebContexts
+
+class GroovyWebContextActionGroup : DefaultActionGroup(
+    "Choose Spring Web Context",
+    true
+) {
+
+    init {
+        templatePresentation.putClientProperty(ActionUtil.SHOW_TEXT_IN_TOOLBAR, true)
+    }
+
+    override fun getChildren(e: AnActionEvent?): Array<out AnAction> {
+        val webContexts = e
+            ?.getData(CommonDataKeys.EDITOR)
+            ?.groovyWebContexts
+            ?: emptyList()
+
+        return buildList {
+            add(Separator.create("Web Application Context"))
+            add(GroovyWebContextAction(GroovyExecConstants.DEFAULT_WEB_CONTEXT))
+            add(Separator.create(""))
+            webContexts.forEach { webContext -> add(GroovyWebContextAction(webContext)) }
+            add(Separator.create())
+            add(ActionManager.getInstance().getAction("yGroovyWebContextsLoad"))
+        }
+            .toTypedArray()
+    }
+
+    override fun update(e: AnActionEvent) {
+        val editor = e.getData(CommonDataKeys.EDITOR) ?: return
+        val activeWebContext = editor.groovyExecContextSettings?.webContext
+            ?: GroovyExecConstants.DEFAULT_WEB_CONTEXT
+
+        e.presentation.text = "Context: $activeWebContext"
+        e.presentation.description = "Web application context"
+        e.presentation.icon = HybrisIcons.Groovy.WEB_CONTEXT_ACTIVE
+    }
+
+    override fun getActionUpdateThread() = ActionUpdateThread.BGT
+}
+